Move table checking to snowflake adapter
